+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 - Update json_schemer to version 2.0
-- Make parameters available at `env[OpenapiFirst::PATH_PARAMS]`, `env[OpenapiFirst::QUERY_PARAMS]`, `env[OpenapiFirst::HEADER_PARAMS]`, `env[OpenapiFirst::COOKIE_PARAMS]` in case you need to access them separately.
+- Breaking: Requires Ruby 3.1 or later
 - Added: Parameters are available at `env[OpenapiFirst::PATH_PARAMS]`, `env[OpenapiFirst::QUERY_PARAMS]`, `env[OpenapiFirst::HEADER_PARAMS]`, `env[OpenapiFirst::COOKIE_PARAMS]` in case you need to access them separately. Merged path and query parameters are still available at `env[OpenapiFirst::PARAMS]`
 - Breaking / Added: ResponseValidation now validates response headers
 - Breaking / Added: RequestValidation now validates cookie, path and header parameters